+++ star/private/ctrls_io.f90
@@ -108,14 +108,16 @@
     TDC_alpha_D, TDC_alpha_R, TDC_alpha_Pt, TDC_alpha_M, &
     TDC_alpha_C, TDC_alpha_S, &
     TDC_alpha_M_use_explicit_mlt_vc_in_momentum_equation, &
-    TDC_use_density_form_for_eddy_viscosity, TDC_adjust_mass_fallback_to_mlt, &
+    TDC_adjust_mass_fallback_to_mlt, &
     TDC_num_innermost_cells_forced_nonturbulent, TDC_num_outermost_cells_forced_nonturbulent, &
     include_mlt_Pturb_in_thermodynamic_gradients, &
     include_mlt_corr_to_TDC, use_TDC_enthalpy_flux_limiter, &
     use_face_reconstruction, &
     TDC_include_eturb_in_energy_equation, &
     use_rsp_form_of_scale_height, include_mlt_in_velocity_time_centering, &
-    TDC_hydro_use_mass_interp_face_values, TDC_hydro_nz, TDC_hydro_nz_outer, TDC_hydro_T_anchor, TDC_hydro_dq_1_factor, &
+    TDC_hydro_use_mass_interp_face_values, TDC_hydro_nz, TDC_hydro_nz_outer, TDC_hydro_nz_inner, &
+    TDC_hydro_nz_T_gradient, &
+    TDC_hydro_T_anchor, TDC_hydro_dq_1_factor, &
 
     ! burn zone eps definitions for use in logs and profiles
     burn_min1, burn_min2, &
@@ -366,7 +368,6 @@
     use_TDC_Y_face_seeded_newton, &
     hydro_matrix_solver, &
     remesh_for_TDC_pulsations_when_load, remesh_for_TDC_pulsations_log_core_zoning, &
-    remesh_for_TDC_pulsations_use_T_gradient, &
     velocity_logT_lower_bound, &
     max_dt_yrs_for_velocity_logT_lower_bound, velocity_tau_lower_bound, velocity_q_upper_bound, &
     use_drag_energy, drag_coefficient, min_q_for_drag, &
@@ -2097,7 +2098,6 @@ s% gradT_excess_max_log_tau_full_off = gradT_excess_max_log_tau_full_off
  s% TDC_alpha_C = TDC_alpha_C
  s% TDC_alpha_S = TDC_alpha_S
  s% TDC_alpha_M_use_explicit_mlt_vc_in_momentum_equation = TDC_alpha_M_use_explicit_mlt_vc_in_momentum_equation
- s% TDC_use_density_form_for_eddy_viscosity = TDC_use_density_form_for_eddy_viscosity
  s% TDC_adjust_mass_fallback_to_mlt = TDC_adjust_mass_fallback_to_mlt
  s% TDC_num_innermost_cells_forced_nonturbulent = TDC_num_innermost_cells_forced_nonturbulent
  s% TDC_num_outermost_cells_forced_nonturbulent = TDC_num_outermost_cells_forced_nonturbulent
@@ -2114,12 +2114,13 @@ s% gradT_excess_max_log_tau_full_off = gradT_excess_max_log_tau_full_off
  s% TDC_hydro_use_mass_interp_face_values = TDC_hydro_use_mass_interp_face_values
  s% TDC_hydro_nz = TDC_hydro_nz
  s% TDC_hydro_nz_outer = TDC_hydro_nz_outer
+ s% TDC_hydro_nz_inner = TDC_hydro_nz_inner
+ s% TDC_hydro_nz_T_gradient = TDC_hydro_nz_T_gradient
  s% TDC_hydro_T_anchor = TDC_hydro_T_anchor
  s% TDC_hydro_dq_1_factor = TDC_hydro_dq_1_factor
 
  s% remesh_for_TDC_pulsations_when_load = remesh_for_TDC_pulsations_when_load
  s% remesh_for_TDC_pulsations_log_core_zoning = remesh_for_TDC_pulsations_log_core_zoning
- s% remesh_for_TDC_pulsations_use_T_gradient = remesh_for_TDC_pulsations_use_T_gradient
 
  s% RSP2_alfap = RSP2_alfap
  s% RSP2_alfad = RSP2_alfad
@@ -3814,7 +3815,6 @@ solver_test_partials_sink_name = s% solver_test_partials_sink_name
  TDC_alpha_C = s% TDC_alpha_C
  TDC_alpha_S = s% TDC_alpha_S
  TDC_alpha_M_use_explicit_mlt_vc_in_momentum_equation = s% TDC_alpha_M_use_explicit_mlt_vc_in_momentum_equation
- TDC_use_density_form_for_eddy_viscosity = s% TDC_use_density_form_for_eddy_viscosity
  TDC_adjust_mass_fallback_to_mlt = s% TDC_adjust_mass_fallback_to_mlt
  TDC_num_innermost_cells_forced_nonturbulent = s% TDC_num_innermost_cells_forced_nonturbulent
  TDC_num_outermost_cells_forced_nonturbulent = s% TDC_num_outermost_cells_forced_nonturbulent
@@ -3831,12 +3831,13 @@ solver_test_partials_sink_name = s% solver_test_partials_sink_name
  TDC_hydro_use_mass_interp_face_values = s% TDC_hydro_use_mass_interp_face_values
  TDC_hydro_nz = s% TDC_hydro_nz
  TDC_hydro_nz_outer = s% TDC_hydro_nz_outer
+ TDC_hydro_nz_inner = s% TDC_hydro_nz_inner
+ TDC_hydro_nz_T_gradient = s% TDC_hydro_nz_T_gradient
  TDC_hydro_T_anchor = s% TDC_hydro_T_anchor
  TDC_hydro_dq_1_factor = s% TDC_hydro_dq_1_factor
 
  remesh_for_TDC_pulsations_when_load = s% remesh_for_TDC_pulsations_when_load
  remesh_for_TDC_pulsations_log_core_zoning = s% remesh_for_TDC_pulsations_log_core_zoning
- remesh_for_TDC_pulsations_use_T_gradient = s% remesh_for_TDC_pulsations_use_T_gradient
 
  RSP2_alfap= s% RSP2_alfap
  RSP2_alfad = s% RSP2_alfad